requirements.txt
attrs==24.2.0
build==1.2.1
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.3.2
cli-exit-tools==1.2.6
click==8.1.7
cronitor==4.7.1
dill==0.3.8
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
gspread-dataframe==4.0.0
humanize==4.10.0
idna==3.8
iniconfig==2.0.0
langdetect==1.0.9
lib-detect-testenv==2.0.8
lingua-language-detector==2.0.2
more-itertools==10.5.0
multiprocess==0.70.16
numpy==2.1.1
oauthlib==3.2.2
packaging==24.2
pandas==2.2.2
platformdirs==4.3.2
pluggy==1.5.0
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyproject_hooks==1.1.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soundchartspy==0.1.3
spotipy==2.24.0
tabulate==0.9.0
timeit-decorator==1.1.1
tqdm==4.66.5
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.3
wrapt==1.16.0
wrapt_timeout_decorator==1.5.1
